+++ example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/repository/EmbeddedDocumentSearchRepository.kt
@@ -1,4 +1,4 @@
-package com.github.inflab.example.spring.data.mongodb.search
+package com.github.inflab.example.spring.data.mongodb.repository
 
 import com.github.inflab.example.spring.data.mongodb.entity.supplies.SaleItem
 import com.github.inflab.example.spring.data.mongodb.entity.supplies.Sales
