+++ example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/repository/ScoreSearchRepository.kt
@@ -1,4 +1,4 @@
-package com.github.inflab.example.spring.data.mongodb.search
+package com.github.inflab.example.spring.data.mongodb.repository
 
 import com.github.inflab.example.spring.data.mongodb.entity.mflix.MovieImdb
 import com.github.inflab.example.spring.data.mongodb.entity.mflix.Movies
